providers/mana: move CQs and work requests to separate files
Re-organize code to have cq.c and wr.c with CQ and WR code, respectively.
